docs: 각 모달 단계는 Index와 함께 위치
