+++ apps/client/src/shared/component/TimeBlockModal/UploadModal.tsx
@@ -6,8 +6,8 @@ import { components } from '@/shared/__generated__/schema';
 import { $api } from '@/shared/api/client';
 import { Modal } from '@/shared/component/Modal';
 import NewFileImportModal from '@/shared/component/NewFileImportModal/NewFileImportModal';
-import BrowseFileModal from '@/shared/component/TimeBlockModal/component/UploadModal/BrowseFile/BrowseFileModal';
-import { flexStyle } from '@/shared/component/TimeBlockModal/component/UploadModal/UploadModal.style';
+import BrowseFileModal from '@/shared/component/TimeBlockModal/component/BrowseFile/BrowseFileModal';
+import { flexStyle } from '@/shared/component/TimeBlockModal/style';
 import { useBlockContext } from '@/shared/hook/common/useBlockContext';
 import { useFunnel } from '@/shared/hook/common/useFunnel';
 import { useInitializeTeamId } from '@/shared/hook/common/useInitializeTeamId';
