--- apps/client/src/shared/component/TimeBlockModal/component/SelectedFileModal/SelectedFileModal.style.ts
@@ -1,15 +1,0 @@
-import { css } from '@emotion/react';
-import { scrollStyle } from '@tiki/ui';
-
-export const fileListWrapperStyle = css({
-  width: '100%',
-  maxHeight: '30rem',
-
-  paddingTop: '2rem',
-
-  overflowY: 'scroll',
-
-  scrollBehavior: 'smooth',
-
-  ...scrollStyle,
-});
